refactor: split content mapper integrations
